requirements.txt
streamlit
google-generativeai
spacy
python-dotenv
spacy download en_core_web_sm
